# Day-1/resources/taskflow-openapi-spec.yml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    TaskFlow is a project management system designed with domain-driven principles.
    
    This API demonstrates business-focused design where URLs speak business language,
    not database language. Notice how endpoints like `/teams/engineering/projects` 
    read like business conversations.
    
    **Key Design Principles:**
    - Domain-driven endpoint design
    - Business language in URLs and responses
    - Professional error handling
    - RESTful resource relationships
    
  version: 1.0.0
  contact:
    name: TaskFlow API Team
    url: https://github.com/your-org/taskflow-api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8001
    description: Local development server
  - url: http://localhost:8001/api/v1
    description: Local development server (versioned)
...
paths:
...
  /tasks/{taskId}:
    get:
      tags: [tasks]
      summary: Get task details
      description: Get detailed information about a specific task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          example: "task_001"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/TaskDetails'

    put:
      tags: [tasks]
      summary: Update task
      description: Update task details and status
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          example: "task_001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
            example:
              status: "in_progress"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/Task'
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          example: "website-redesign"
        name:
          type: string
          example: "Website Redesign"
        description:
          type: string
          example: "Complete overhaul of company website"
        status:
          type: string
          enum: [planning, in_progress, completed, on_hold]
          example: "in_progress"
        teamId:
          type: string
          example: "engineering"
        priority:
          type: string
          enum: [low, medium, high, critical]
          example: "high"
        tasksCount:
          type: integer
          example: 12
        completedTasks:
          type: integer
          example: 8
        startDate:
          type: string
          format: date
          example: "2024-01-01"
        dueDate:
          type: string
          format: date
          example: "2024-03-01"
        createdAt:
          type: string
          format: date-time
...
    Task:
      type: object
      properties:
        id:
          type: string
          example: "task_001"
        title:
          type: string
          example: "Implement user authentication"
        description:
          type: string
          example: "Add JWT-based login system with role management"
        status:
          type: string
          enum: [todo, in_progress, completed, blocked]
          example: "todo"
        priority:
          type: string
          enum: [low, medium, high, critical]
          example: "high"
        projectId:
          type: string
          example: "website-redesign"
        assignedTo:
          type: string
          nullable: true
          example: "jane.smith"
        assignedBy:
          type: string
          nullable: true
          example: "john.doe"
        estimatedHours:
          type: integer
          nullable: true
          example: 8
        actualHours:
          type: integer
          nullable: true
          example: 6
        createdAt:
          type: string
          format: date-time
        assignedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
    TaskDetails:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
            project:
              $ref: '#/components/schemas/Project'
...
    Comment:
      type: object
      properties:
        id:
          type: string
          example: "comment_001"
        taskId:
          type: string
          example: "task_001"
        comment:
          type: string
          example: "Implemented basic JWT middleware, next up is role validation"
        author:
          type: string
          example: "jane.smith"
        createdAt:
          type: string
          format: date-time
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, completed, blocked]
        priority:
          type: string
          enum: [low, medium, high, critical]
        estimatedHours:
          type: integer
        actualHours:
          type: integer
        notes:
          type: string
...